requirements.txt
Flask==2.1.3
Flask-Cors==3.0.10
langchain==0.1.0
pinecone==6.0.2
python-dotenv==1.0.0
openai==1.2.3